Move utilities to a package to keep things clean
